Working query for contributed repos
